OSS接口能力扩展

+++ web-server/src/main/java/com/shigu/seller/actions/GoodsFileAction.java
@@ -94,7 +94,7 @@ public class GoodsFileAction {
     @RequestMapping("seller/getGlGoodslist")
     @ResponseBody
     public JSONObject getGlGoodslist(Integer page,String fileId,HttpSession session) throws JsonErrException {
-        if(org.apache.commons.lang3.StringUtils.isEmpty(fileId)){
+        if (!goodsFileService.checkFileId(fileId)) {
             throw new JsonErrException("key信息异常");
         }
         if (page == null) {
@@ -112,7 +112,10 @@ public class GoodsFileAction {
      */
     @RequestMapping("seller/readGlGoodsJson")
     @ResponseBody
-    public JSONObject selGoodsFileByFile(String fileKey,HttpSession session) {
+    public JSONObject selGoodsFileByFile(String fileKey,HttpSession session) throws JsonErrException {
+        if (!goodsFileService.checkFileId(fileKey)) {
+            throw new JsonErrException("key信息异常");
+        }
         List<ItemShowBlock> goodsFilesList = goodsFileService.selGoodsFileByFile(logshop(session).getShopId(),fileKey);
         return JsonResponseUtil.success().element("goods", goodsFilesList);
     }
@@ -126,6 +129,10 @@ public class GoodsFileAction {
     @RequestMapping("seller/glGoodsJson")
     @ResponseBody
     public JSONObject saveGoodsFile(String fileId, String goodsIds,HttpSession session) throws JsonErrException {
+        if (!goodsFileService.checkFileId(fileId)) {
+            throw new JsonErrException("key信息异常");
+        }
+
         ShopSession shop=logshop(session);
         goodsFileService.saveGoodsFile(fileId,shop.getShopId(), goodsIds,shop.getWebSite());
         return JsonResponseUtil.success();
@@ -139,6 +146,9 @@ public class GoodsFileAction {
     @RequestMapping("seller/cacelGlJson")
     @ResponseBody
     public JSONObject delGoodsFile(Long goodsId,String fileId,HttpSession session) throws JsonErrException {
+        if (!goodsFileService.checkFileId(fileId)) {
+            throw new JsonErrException("key信息异常");
+        }
         ShopSession shop=logshop(session);
         goodsFileService.delGoodsfile(goodsId,shop.getShopId(),fileId,shop.getWebSite());
         return JsonResponseUtil.success();
@@ -153,7 +163,10 @@ public class GoodsFileAction {
      */
     @RequestMapping("seller/renameFile")
     @ResponseBody
-    public JSONObject renameFile(String fileId, String fileType, String newName,HttpSession session) {
+    public JSONObject renameFile(String fileId, String fileType, String newName,HttpSession session) throws JsonErrException  {
+        if (!goodsFileService.checkFileId(fileId)) {
+            throw new JsonErrException("key信息异常");
+        }
         goodsFileService.rename(logshop(session).getShopId(),fileId, fileType, newName);
 
         return JsonResponseUtil.success();
@@ -167,7 +180,10 @@ public class GoodsFileAction {
      */
     @RequestMapping("seller/moveFile")
     @ResponseBody
-    public JSONObject moveFile(String fileId, String targetFolderId,HttpSession session) {
+    public JSONObject moveFile(String fileId, String targetFolderId,HttpSession session) throws JsonErrException {
+        if (!goodsFileService.checkFileId(fileId)) {
+            throw new JsonErrException("key信息异常");
+        }
         goodsFileService.moveFile(logshop(session).getShopId(),fileId, targetFolderId);
         return JsonResponseUtil.success();
     }
@@ -201,6 +217,9 @@ public class GoodsFileAction {
     @RequestMapping("seller/deleteFile")
     @ResponseBody
     public JSONObject deleteFile(String fileId, String fileType,HttpSession session) throws JsonErrException {
+        if (!goodsFileService.checkFileId(fileId)) {
+            throw new JsonErrException("key信息异常");
+        }
         if(goodsFileService.deleteFile(logshop(session).getShopId(),fileId, fileType)){
             return JsonResponseUtil.success();
         }
@@ -215,7 +234,10 @@ public class GoodsFileAction {
      */
     @RequestMapping("seller/getFileList")
     @ResponseBody
-    public JSONObject getFileList(String fileId,HttpSession session) {
+    public JSONObject getFileList(String fileId,HttpSession session) throws JsonErrException {
+        if (!goodsFileService.checkFileId(fileId)) {
+            throw new JsonErrException("key信息异常");
+        }
         List<GoodsFileVO> files = goodsFileService.selFilesByFileId(logshop(session).getShopId(),fileId);
         JSONObject obj= JsonResponseUtil.success();
         if(org.apache.commons.lang3.StringUtils.isNotEmpty(fileId)&&fileId.contains("/")){
@@ -253,6 +275,9 @@ public class GoodsFileAction {
     @RequestMapping("seller/noticeUploadFile")
     @ResponseBody
     public JSONObject noticeUploadFile(String fileId,HttpSession session) throws JsonErrException {
+        if (!goodsFileService.checkFileId(fileId)) {
+            throw new JsonErrException("key信息异常");
+        }
         GoodsFileVO vo=goodsFileService.uploadFile(logshop(session).getShopId(),fileId);
         return JSONObject.fromObject(vo).element("result","success");
     }
